Nft suggestions (#116)
* nft suggestions

+++ packages/tokens/non-fungible/src/extensions/burnable/mod.rs
@@ -34,13 +34,13 @@ pub trait NonFungibleBurnable {
     /// # Events
     ///
     /// * topics - `["burn", from: Address]`
-    /// * data - `[token_id: u128]`
+    /// * data - `[token_id: u32]`
     ///
     /// # Notes
     ///
     /// We recommend using [`crate::burnable::burn()`] when implementing this
     /// function.
-    fn burn(e: &Env, from: Address, token_id: u128);
+    fn burn(e: &Env, from: Address, token_id: u32);
 
     /// Destroys the `token_id` from `account`, by using `spender`s approval.
     ///
@@ -64,13 +64,13 @@ pub trait NonFungibleBurnable {
     /// # Events
     ///
     /// * topics - `["burn", from: Address]`
-    /// * data - `[token_id: u128]`
+    /// * data - `[token_id: u32]`
     ///
     /// # Notes
     ///
     /// We recommend using [`crate::burnable::burn_from()`] when implementing
     /// this function.
-    fn burn_from(e: &Env, spender: Address, from: Address, token_id: u128);
+    fn burn_from(e: &Env, spender: Address, from: Address, token_id: u32);
 }
 
 // ################## EVENTS ##################
@@ -86,8 +86,8 @@ pub trait NonFungibleBurnable {
 /// # Events
 ///
 /// * topics - `["burn", from: Address]`
-/// * data - `[token_id: u128]`
-pub fn emit_burn(e: &Env, from: &Address, token_id: u128) {
+/// * data - `[token_id: u32]`
+pub fn emit_burn(e: &Env, from: &Address, token_id: u32) {
     let topics = (symbol_short!("burn"), from);
     e.events().publish(topics, token_id)
 }
