+++ packages/tokens/non-fungible/src/extensions/burnable/storage.rs
@@ -21,12 +21,12 @@ use crate::{
 /// # Events
 ///
 /// * topics - `["burn", from: Address]`
-/// * data - `[token_id: u128]`
+/// * data - `[token_id: u32]`
 ///
 /// # Notes
 ///
 /// Authorization for `from` is required.
-pub fn burn(e: &Env, from: &Address, token_id: u128) {
+pub fn burn(e: &Env, from: &Address, token_id: u32) {
     from.require_auth();
     update(e, Some(from), None, token_id);
     emit_burn(e, from, token_id);
@@ -51,12 +51,12 @@ pub fn burn(e: &Env, from: &Address, token_id: u128) {
 /// # Events
 ///
 /// * topics - `["burn", from: Address]`
-/// * data - `[token_id: u128]`
+/// * data - `[token_id: u32]`
 ///
 /// # Notes
 ///
 /// Authorization for `spender` is required.
-pub fn burn_from(e: &Env, spender: &Address, from: &Address, token_id: u128) {
+pub fn burn_from(e: &Env, spender: &Address, from: &Address, token_id: u32) {
     spender.require_auth();
     check_spender_approval(e, spender, from, token_id);
     update(e, Some(from), None, token_id);
